[performance optimization] tsdb: Merge two loops in ‘delete’ function