@csmarchbanks 2020-03-09
Do you have benchmarks showing a performance improvement? This change will always alloce a slice even in the case where nothing is to be deleted which could cause worse performance.