@liguozhong 2020-03-10
> Do you have benchmarks showing a performance improvement? This change will always alloce a slice even in the case where nothing is to be deleted which could cause worse performance.

yes. I run the benchmark and confirmed that it can double the performance.
![image](https://user-images.githubusercontent.com/9583245/76315346-782d8900-6313-11ea-87ff-c475301e018f.png)
(this pr)
![image](https://user-images.githubusercontent.com/9583245/76315418-998e7500-6313-11ea-819f-596ee93065b0.png)
(old delete )