@liguozhong 2020-03-11
> It will probably be hard to tell in a prombench if this makes any difference as I think it would only happen during compaction where a lot of cpu intensive things are happening.
> 
> To merge this I would like to see benchmarks added to this MR that can be verified, when I ran the benchmark that seems to have been deleted from this thread I actually found the new code to be slower.
> 
> ```
> benchmark                         old ns/op     new ns/op     delta
> BenchmarkMemPostings_Delete-8     94931304      105818435     +11.47%
> ```
OK, I'll try