@gouthamve 2020-08-31
Hi @liguozhong, thanks for submitting the PR, seems useful, but with perf its always tricky. Did you have new benchmarks for this?