@liguozhong 2020-09-01
> Hi @liguozhong, thanks for submitting the PR, seems useful, but with perf its always tricky. Did you have new benchmarks for this?

My benchmark shows that this PR makes the performance worse, this is not a good idea